Several errors corrected
Timing of all agents must be given by the tick that contains the current
tick. It is more correct that every agent containing a private value of
the time running.

JSONArray of a ArrayList is now free of running errors, because there is
not a direct way to introduce ArrayList in the constructor of a
JSONArray, so the ArrayList is transformed as a new array that its is
finally given to the constructor of a JSONArray.

Ratio for each carAgent is provided as an argument in the
EventManagerAgent behaviour that throws the carAgent (and it must be
read from the configuration file of every car).

There is for the carAgent, the currentTick received from the tick and
the tini that stores the initial time in which an agent is driving in a
new segment of its route.

The ids of the car sending and receiving TrafficData is now introduced.
Maybe it will be used later for some calculation on measure timing of
the same car in the same segment or nextSegment?