Major Changes on Dijkstra and on ways to compute weights
Dijkstra drawn from TrafficMap and arranged in a new package that uses
the factory and strategy design patters to get a flexible way to
introduce new methods to compute routes without changing the code of
CarAgent. When this happens changes has to be done in Method enumeration
class, RouteFactory class and write the new class that contains only a
public method for computing the weights. This new class must extend the
Route abstract class.